electrafy: upstream-compat fixes for smoke test on upstream/main
Four small fixes uncovered by running pytest against upstream/main:

* model.py: cast species to long before zval_lookup indexing. Recent
  metatrain's systems_to_batch hands species back as the model dtype
  (float32) for downstream attention plumbing; previous fork branches
  kept it as int. Defensive cast survives both layouts.

* modules/loss.py: vendor TensorMapNMAELoss (LossInterface subclass)
  from origin/electrafi-streaming-triton. Same math as the inline
  NMAELoss but follows the metatrain LossAggregator API (operates on
  Dict[str, TensorMap], implements compute(predictions, targets,
  extra_data)).

* __init__.py: monkey-patch metatrain.utils.loss.create_loss to
  intercept the 'nmae' type key before LossType.from_key raises.
  Upstream's LossType is a closed Enum so we cannot register new types
  the standard way; wrapping the factory at electrafy-import time is
  idempotent and keeps all the changes inside experimental/electrafy/.
  Enables LossSpecification(type='nmae') in standard metatrain configs.

* pyproject.toml: extend [tool.pytest.ini_options].filterwarnings to
  ignore three torch-internal UserWarnings/DeprecationWarnings that fire
  on every test collection because electrafy/modules/fourier_density.py
  module-level torch.compile pulls in torch._inductor and torch.utils.mkldnn:
    - torch.jit.script_method deprecation (from torch.utils.mkldnn)
    - Torchinductor complex-operator codegen fallback
    - non-leaf Tensor .grad access (from torch._dynamo.safe_has_grad
      wrapping covs, which is a non-leaf chol@chol product)
  None of these are real bugs; they were just promoted to errors by
  upstream's filterwarnings='error' default.

Smoke test (job 3368485): 66 passed, 21 skipped (skips are fixture-
dependent: real CHGCAR cache files, etc.), 0 failed.

+++ src/metatrain/experimental/electrafy/model.py
@@ -359,7 +359,10 @@ class ELECTRAFY(ModelInterface[ModelHypers]):
 
         # ---- Stages 3-4: Per-system density ----
         positions_all = torch.cat([sys.positions for sys in systems], dim=0)
-        n_gaussians_per_atom = self.zval_lookup[species] * self.gaussians_per_electron
+        # ``species`` may arrive as float (recent metatrain casts System.types
+        # to the model dtype for downstream attention plumbing); cast back to
+        # long for the integer-index lookup.
+        n_gaussians_per_atom = self.zval_lookup[species.long()] * self.gaussians_per_electron
 
         # Resolve per-system output grid shapes (paper convention: native
         # NGXF/NGYF/NGZF). When ``set_override_grid_shapes`` has been called,
